django/requirements.txt
APScheduler==3.6.0
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
Django==2.2.4
django-bootstrap4==0.0.8
django-fontawesome-5==1.0.16
django-humanize==0.1.2
djangorestframework==3.10.0
gunicorn==19.9.0
humanize==0.5.1
idna==2.8
inflection==0.3.1
Jinja2==2.10.1
logstash-formatter==0.5.17
Markdown==3.1.1
MarkupSafe==1.1.1
pytz==2019.1
PyYAML==5.1.1
six==1.12.0
tzlocal==1.5.1
whitenoise==4.1.3
